+++ src/egon/data/datasets/electricity_demand_timeseries/hh_buildings.py
@@ -13,7 +13,6 @@ The resulting data is stored in separate tables
 * `demand.egon_household_electricity_profile_of_buildings`:
     Mapping of demand timeseries and buildings including cell_id, building
     area and peak load
-* 'demand.egon
 
 Both tables are created within :func:`map_houseprofiles_to_buildings`.
 
@@ -53,11 +52,11 @@ areas not considered in census 2011.
 The assignment of household demand timeseries to buildings takes place at cell
 level. Within each cell a pool of profiles exists, produced by the 'HH Demand"
 module. These profiles are randomly assigned to a filtered list of OSM buildings
-within this cell. The assignment takes place with replacement of every building
-after drew. Therefore multiple profiles can be assigned to one building, making
-it a multi-household building. Hence, every profile is assigned to a building
-but not every building needs to have a profile assigned as not all are
-residential.
+within this cell. Every profile is assigned to a building and every building
+get a profile assigned if there is enough households by the census data. If
+there are more profiles then buildings, all additional profiles are randomly
+assigned. Therefore multiple profiles can be assigned to one building, making
+it a multi-household building.
 
 
 **What are central assumptions during the data processing?**
@@ -78,6 +77,27 @@ integer. Ceiling is applied to avoid rounding to amount of 0 buildings.
 * As this datasets is a cascade after profile assignement at census cells
 also check drawbacks and limitations in hh_profiles.py.
 
+Example Query
+-----
+
+* Get a list with number of houses, households and household types per census cell
+
+    SELECT t1.cell_id, building_count, hh_count, hh_types
+        FROM(
+            SELECT cell_id, Count(distinct(building_id)) as building_count,
+            count(profile_id) as hh_count
+                FROM demand.egon_household_electricity_profile_of_buildings
+            Group By cell_id
+        ) as t1
+    FULL OUTER JOIN(
+        SELECT cell_id, array_agg(array[cast(hh_10types as char),
+         hh_type]) as hh_types
+        FROM society.egon_destatis_zensus_household_per_ha_refined
+        GROUP BY cell_id
+        ) as t2
+    ON t1.cell_id = t2.cell_id
+
+
 Notes
 -----
 
@@ -86,6 +106,7 @@ is made in ... the content of this module docstring needs to be moved to
 docs attribute of the respective dataset class.
 """
 from functools import partial
+import random
 
 from geoalchemy2 import Geometry
 from shapely.geometry import Point
@@ -398,7 +419,8 @@ def generate_mapping_table(
     All hh demand profiles are randomly assigned to buildings within the same
     cencus cell.
 
-    * profiles > buildings: buildings have multiple profiles
+    * profiles > buildings: buildings can have multiple profiles but every
+        building gets at least one profile
     * profiles < buildings: not every building gets a profile
 
 
@@ -415,6 +437,18 @@ def generate_mapping_table(
         Table with mapping of profile ids to buildings with OSM ids
 
     """
+
+    def create_pool(buildings, profiles):
+
+        if profiles > buildings:
+            surplus = profiles - buildings
+            surplus = rng.integers(0, buildings, surplus)
+            pool = list(range(buildings)) + list(surplus)
+        else:
+            pool = list(range(buildings))
+        result = random.sample(population=pool, k=profiles)
+
+        return result
 
     # group oms_ids by census cells and aggregate to list
     osm_ids_per_cell = (
@@ -457,10 +491,12 @@ def generate_mapping_table(
     )
 
     # map profiles randomly per cell
+    # if profiles > buildings, every building will get at least one profile
     rng = np.random.default_rng(RANDOM_SEED)
+    random.seed(RANDOM_SEED)
     mapping_profiles_to_buildings = pd.Series(
         [
-            rng.integers(0, buildings, profiles)
+            create_pool(buildings, profiles)
             for buildings, profiles in zip(
                 number_profiles_and_buildings_reduced["building_ids"].values,
                 number_profiles_and_buildings_reduced[
@@ -470,6 +506,7 @@ def generate_mapping_table(
         ],
         index=number_profiles_and_buildings_reduced.index,
     )
+
     # unnest building assignement per cell
     mapping_profiles_to_buildings = (
         mapping_profiles_to_buildings.rename("building")
@@ -785,7 +822,7 @@ def map_houseprofiles_to_buildings():
 setup = partial(
     Dataset,
     name="Demand_Building_Assignment",
-    version="0.0.0",
+    version="0.0.1",
     dependencies=[],
     tasks=(map_houseprofiles_to_buildings, get_building_peak_loads),
 )